Refactor: 도메인 타입 수정
Request, Response 폴더 추가
이름 및 파일에 type 포함으로 변경
선택 요소 ? 추가
Request 타입 이름 변경
