+++ src/Types/ChannelType.ts
@@ -1,5 +1,5 @@
-export interface Channel {
-  authRequired: boolean; // 사용되지 않음
+export interface ChannelType {
+  authRequired?: boolean; // 사용되지 않음
   posts: string[];
   _id: string;
   name: string;
