+++ src/Types/ConversationType.ts
@@ -1,0 +1,10 @@
+import { UserType } from './UserType';
+
+export interface ConversationType {
+  _id: string[];
+  message: string;
+  sender: UserType;
+  receiver: UserType;
+  seen: boolean;
+  createdAt: string;
+}
