+++ src/Types/FollowType.ts
@@ -1,4 +1,4 @@
-export interface Follow {
+export interface FollowType {
   _id: string;
   user: string; // 사용자 id
   follower: string; // 사용자 id
