--- src/Types/Interface/Conversation.ts
@@ -1,10 +1,0 @@
-import { User } from './User';
-
-export interface Conversation {
-  _id: string[];
-  message: string;
-  sender: User;
-  receiver: User;
-  seen: boolean;
-  createdAt: string;
-}
