--- src/Types/Interface/Notification.ts
@@ -1,21 +1,0 @@
-import { User } from './User';
-
-export interface Notification {
-  seen: boolean;
-  _id: string;
-  author: User;
-  user: User | string;
-  post: string | null; // 포스트 id
-  follow?: string; // 사용자 id
-  comment?: Comment;
-  message?: string; // 메시지 id
-  createdAt: string;
-  updatedAt: string;
-}
-
-export interface CreateNotificationRequest {
-  notificationType: 'COMMENT' | 'FOLLOW' | 'LIKE' | 'MESSAGE';
-  notificationTypeId: string;
-  userId: string;
-  postId: string | null;
-}
