--- src/Types/Interface/Post.ts
@@ -1,23 +1,0 @@
-import { Channel } from './Channel';
-import { Comment } from './Comment';
-import { Like } from './Like';
-import { User } from './User';
-
-export interface Post {
-  likes: Like[];
-  comments: Comment[];
-  _id: string;
-  image?: string;
-  imagePublicId: string;
-  title: string;
-  channel: Channel;
-  author: User;
-  createdAt: string;
-  updatedAt: string;
-}
-
-export interface CreatePostRequest {
-  title: string;
-  image: BinaryData | null;
-  channelId: string;
-}
