+++ src/Types/Response/index.ts
@@ -1,0 +1,6 @@
+import { UserType } from '../UserType';
+
+export interface UserResponse {
+  user: UserType;
+  token: string;
+}
